% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateObject.R
\name{CreateVisObject}
\alias{CreateVisObject}
\title{Title}
\usage{
CreateVisObject(
  original_matrix = NULL,
  input_path = NULL,
  min.cell = 0,
  min.gene = 0
)
}
\arguments{
\item{original_matrix}{input original expression matrix}

\item{input_path}{input directory path which contain three major output and original matrix.}

\item{min.gene}{}
}
\value{

}
\description{
Title
}
