% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/RunLTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,LTMG-method}
\title{RunLTMG}
\usage{
RunLTMG(object, ...)

.RunLTMG(object, Gene_use = NULL, seed = 123, k = 5)

\S4method{RunLTMG}{LTMG}(object, Gene_use = NULL, seed = 123, k = 5)
}
\arguments{
\item{Gene_use}{using X numebr of top variant gene. input a number, recommend 2000.}
}
\value{

}
\description{
RunLTMG
}
