% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/InputData.R
\name{WriteSparse}
\alias{WriteSparse}
\alias{writesparse}
\alias{WriteSparse,LTMG-method}
\title{Title}
\usage{
WriteSparse(object, ...)

writesparse(object = NULL)

\S4method{WriteSparse}{LTMG}(object = NULL)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Title
}
