% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/InputData.R
\name{WriteSparse}
\alias{WriteSparse}
\alias{writesparse}
\alias{WriteSparse,LTMG-method}
\title{Title}
\usage{
WriteSparse(object, ...)

writesparse(object = NULL, path = "./", gene.name = TRUE, cell.name = FALSE)

\S4method{WriteSparse}{LTMG}(object = NULL, path = "./", gene.name = TRUE, cell.name = FALSE)
}
\arguments{
\item{path}{output path}

\item{gene.name}{whether output gene name for sparse matrix}

\item{cell.name}{whether output cell name for sparse matrix}
}
\value{

}
\description{
Title
}
