% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/InputData.R
\name{write.sparse}
\alias{write.sparse}
\alias{WriteSparse}
\alias{write.sparse,LTMG-method}
\title{write sparse matrix}
\usage{
write.sparse(object)

write.sparse(object)

\S4method{write.sparse}{LTMG}(object)
}
\description{
write sparse matrix
}
